# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Get a user by its id
      operationId: getUser
      parameters:
      - name: id
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: User not found
          content: {}
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid id supplied
        "401":
          description: User not authorized
    put:
      tags:
      - user-rest-controller
      summary: Update a user by its id
      operationId: replaceUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: User updated correctly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: User not updated
        "403":
          description: User not authorized
  /api/users/{id}/image:
    get:
      tags:
      - user-rest-controller
      summary: Gets the image of a user by its id
      operationId: getUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "User not found, user image not found or doesn't have permission\
            \ to access it"
          content: {}
        "200":
          description: Found the user image
        "400":
          description: Invalid id supplied
        "401":
          description: User not authorized
    put:
      tags:
      - user-rest-controller
      summary: Updates the image of a user
      operationId: replaceUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: User not found
          content: {}
        "201":
          description: Image created correctly
        "204":
          description: Image updated correctly
        "400":
          description: Bad request
        "401":
          description: User not authorized
        "403":
          description: User not authorized
    post:
      tags:
      - user-rest-controller
      summary: Registers the image of a user
      operationId: createUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: User not found
          content: {}
        "201":
          description: Image created correctly
        "400":
          description: Bad request
        "401":
          description: User not authorized
        "403":
          description: User not authorized
...
  /api/trainings/{trainingId}/image:
    get:
      tags:
      - training-rest-controller
      summary: Retrieve training image
      operationId: getTrainingImage
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training image not found
          content: {}
        "200":
          description: Training image found
        "500":
          description: Internal server error
    put:
      tags:
      - training-rest-controller
      summary: Replace training image
      operationId: replaceTrainingImage
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Not Found
        "204":
          description: Training image replaced
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "500":
          description: Internal server error
    post:
      tags:
      - training-rest-controller
      summary: Create training image
      operationId: createTrainingImage
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "404":
          description: Not Found
        "201":
          description: Training image created
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "500":
          description: Internal server error
    delete:
      tags:
      - training-rest-controller
      summary: Delete training image
      operationId: deletePostImage
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training image not found
          content: {}
        "204":
          description: Training image deleted
        "500":
          description: Internal server error
...
  /api/users/:
    post:
      tags:
      - user-rest-controller
      summary: Registers a new user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: User registered correctly
        "400":
          description: "Bad request, maybe one of the user attributes is missing or\
            \ the type is not valid"
        "409":
          description: User already exists
...
  /api/users/reportedComments:
    get:
      tags:
      - user-rest-controller
      operationId: getReportedComments
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
  /api/users/me:
    get:
      tags:
      - user-rest-controller
      summary: Gets the logged user
      operationId: me
      responses:
        "404":
          description: Not Found
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: User not authorized
  /api/users/all:
    get:
      tags:
      - user-rest-controller
      summary: Get all users
      operationId: getUsers
      responses:
        "404":
          description: Users not found
          content: {}
        "200":
          description: Found all users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: User not authorized
...
components:
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        email:
          type: string
    Nutrition:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        image:
          type: boolean
        goal:
          type: string
        calories:
          type: integer
          format: int32
        comments:
          type: array
          items:
            $ref: "#/components/schemas/NutritionComment"
        user:
          $ref: "#/components/schemas/User"
        nutritionComments:
          type: array
          items:
            $ref: "#/components/schemas/NutritionComment"
          writeOnly: true
    NutritionComment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        name:
          type: string
        isNotified:
          type: boolean
        nutrition:
          $ref: "#/components/schemas/Nutrition"
        user:
          $ref: "#/components/schemas/User"
    Training:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        duration:
          type: integer
          format: int32
        intensity:
          type: string
        description:
          type: string
        image:
          type: boolean
        goal:
          type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/TrainingComment"
        user:
          $ref: "#/components/schemas/User"
        trainingComments:
          type: array
          items:
            $ref: "#/components/schemas/TrainingComment"
          writeOnly: true
        img:
          type: string
    TrainingComment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        name:
          type: string
        isNotified:
          type: boolean
        training:
          $ref: "#/components/schemas/Training"
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        trainingComments:
          type: array
          items:
            $ref: "#/components/schemas/TrainingComment"
        nutritionComments:
          type: array
          items:
            $ref: "#/components/schemas/NutritionComment"
        trainings:
          type: array
          items:
            $ref: "#/components/schemas/Training"
        nutritions:
          type: array
          items:
            $ref: "#/components/schemas/Nutrition"
...